chore: removed support for unused Prometheus and ArgoCd APIs
