+++ internal/sql/repository/chartConfig/PipelineOverrideRepository.go
@@ -19,7 +19,7 @@ package chartConfig
 
 import (
 	"github.com/devtron-labs/devtron/api/bean"
-	"github.com/devtron-labs/devtron/client/argocdServer/application"
+	bean2 "github.com/devtron-labs/devtron/client/argocdServer/bean"
 	"github.com/devtron-labs/devtron/internal/sql/models"
 	"github.com/devtron-labs/devtron/internal/sql/repository"
 	"github.com/devtron-labs/devtron/internal/sql/repository/pipelineConfig"
@@ -231,7 +231,7 @@ func (impl PipelineOverrideRepositoryImpl) FetchHelmTypePipelineOverridesForStat
 		Join("inner join cd_workflow cdwf on cdwf.pipeline_id = p.id").
 		Join("inner join cd_workflow_runner cdwfr on cdwfr.cd_workflow_id = cdwf.id").
 		Where("p.deployment_app_type = ?", util.PIPELINE_DEPLOYMENT_TYPE_HELM).
-		Where("cdwfr.status not in (?)", pg.In([]string{application.Degraded, application.HIBERNATING, application.Healthy, "Failed", "Aborted"})).
+		Where("cdwfr.status not in (?)", pg.In([]string{bean2.Degraded, bean2.HIBERNATING, bean2.Healthy, "Failed", "Aborted"})).
 		Where("cdwfr.workflow_type = ?", bean.CD_WORKFLOW_TYPE_DEPLOY).
 		Where("p.deleted = ?", false).
 		Select()
